Initial vesion v0.1
In this initial version the Arduino circuit measures the environment
light and send it by serial to the C# application which read it every 10
seconds and use it to set the monitor brightness and contrast.
A trackbar is included in the Form to permit the user to manually manage
the monitor brightness.